[POO]Exercício 02 classe abstrata.

+++ src/br/com/geekuniversity/secao12/exercitando/exercicio02/Pagamento.java
@@ -10,6 +10,6 @@ public abstract class Pagamento {
         this.titular = titular;
     }
 
-    public abstract int processarPagamento();
+    public abstract double processarPagamento();
 
 }
